+++ src/main/java/com/metafour/multitenancy/IndexController.java
@@ -3,6 +3,12 @@ package com.metafour.multitenancy;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+/**
+ * Simple index controller.
+ * 
+ * @author Imtiaz Rahi
+ * @since 2017-08-25
+ */
 @Controller
 @RequestMapping("/")
 public class IndexController {