+++ src/main/java/com/metafour/multitenancy/bean/EmployeeRepository.java
@@ -2,6 +2,13 @@ package com.metafour.multitenancy.bean;
 
 import org.springframework.data.repository.CrudRepository;
 
+/**
+ * Employee repository for CRUD operations.
+ * 
+ * @author Imtiaz Rahi
+ * @since 2017-08-25
+ * @see Employee
+ */
 public interface EmployeeRepository extends CrudRepository<Employee, Long> {
 
 }