+++ src/main/java/com/metafour/multitenancy/config/DataSourceConfig.java
@@ -11,6 +11,23 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
 
+/**
+ * Inject multiple JDBC data sources as Spring beans into IOC. <br>
+ * Database configurations are taken from {@code multitenant.properties} file and structured using {@link MultitenancyProperties}.
+ * 
+ * <pre>
+ * <code>
+ * spring.multitenancy.datasource1.url = jdbc:postgresql://localhost:5432/tenant_1
+ * spring.multitenancy.datasource1.username = username
+ * spring.multitenancy.datasource1.password = password
+ * spring.multitenancy.datasource1.driver-class-name = org.postgresql.Driver
+ * </code>
+ * </pre>
+ * 
+ * @author Imtiaz Rahi
+ * @since 2017-08-25
+ * @see MultitenancyProperties
+ */
 @EnableConfigurationProperties(MultitenancyProperties.class)
 @Configuration
 @PropertySource("classpath:multitenant.properties")
@@ -22,22 +39,22 @@ public class DataSourceConfig {
 	@Bean(name = { "dataSource", "dataSource1" })
 	@ConfigurationProperties(prefix = "spring.multitenancy.datasource1")
 	public DataSource dataSource1() {
-		return getDS(this.multitenancyProperties.getDatasource1());
+		return buildDS(this.multitenancyProperties.getDatasource1());
 	}
 
 	@Bean(name = "dataSource2")
 	@ConfigurationProperties(prefix = "spring.multitenancy.datasource2")
 	public DataSource dataSource2() {
-		return getDS(this.multitenancyProperties.getDatasource2());
+		return buildDS(this.multitenancyProperties.getDatasource2());
 	}
 
 	@Bean(name = "dataSource3")
 	@ConfigurationProperties(prefix = "spring.multitenancy.datasource3")
 	public DataSource dataSource3() {
-		return getDS(this.multitenancyProperties.getDatasource3());
+		return buildDS(this.multitenancyProperties.getDatasource3());
 	}
 
-	private DataSource getDS(DataSourceProperties dsprop) {
+	private DataSource buildDS(DataSourceProperties dsprop) {
 		DataSourceBuilder factory = DataSourceBuilder
 				.create(dsprop.getClassLoader())
 				.driverClassName(dsprop.getDriverClassName())