+++ src/main/java/com/metafour/multitenancy/impl/DataSourceBasedMultiTenantConnectionProviderImpl.java
@@ -1,4 +1,4 @@
-package com.metafour.multitenancy;
+package com.metafour.multitenancy.impl;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -10,10 +10,18 @@ import org.hibernate.engine.jdbc.connections.spi.AbstractDataSourceBasedMultiTen
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
+import com.metafour.multitenancy.config.MultiTenancyJpaConfiguration;
+
+/**
+ * Database connection provider implementation for multitenant scenario.
+ *  
+ * @author Imtiaz Rahi
+ * @since 2017-08-25
+ * @see MultiTenancyJpaConfiguration
+ */
 @Component
 public class DataSourceBasedMultiTenantConnectionProviderImpl extends AbstractDataSourceBasedMultiTenantConnectionProviderImpl {
 	private static final long serialVersionUID = 8168907057647334460L;
-	private static final String DEFAULT_TENANT_ID = "tenant_1";
 
 	@Autowired
 	private DataSource dataSource1;
@@ -36,7 +44,7 @@ public class DataSourceBasedMultiTenantConnectionProviderImpl extends AbstractDa
 
 	@Override
 	protected DataSource selectAnyDataSource() {
-		return map.get(DEFAULT_TENANT_ID);
+		return map.get(TenantContext.DEFAULT_TENANT_ID);
 	}
 
 	@Override