+++ src/main/java/com/metafour/multitenancy/impl/package-info.java
@@ -1,0 +1,14 @@
+/**
+ * Copyright (c) 2017 Imtiaz Rahi.
+ * 
+ * This software may be modified and distributed under the terms
+ * of the MIT license.  See the LICENSE file for details.
+ */
+
+/**
+ * Multitenancy implementation details with custom classes.
+ * 
+ * @author Imtiaz Rahi
+ * @since 2017-08-28
+ */
+package com.metafour.multitenancy.impl;